+++ evm-adapters/src/sputnik/mod.rs
@@ -2,7 +2,7 @@ mod evm;
 pub use evm::*;
 
 mod forked_backend;
-pub use forked_backend::{new_shared_cache, ForkMemoryBackend, SharedBackend, SharedCache};
+pub use forked_backend::*;
 
 pub mod cheatcodes;
 pub mod state;
